Use a more robust IF check in the formulas
Fixes #409
In the formulas in columns D, F and G the IF conditions are changed
from E9=0 to A9="", to be able to handle very short segments where the
horizontal distance rounds down to 0.
In the formulas in column K, the IF condition is removed entirely,
because this condition isn't required for the correct behaviour, and it
caused the planned time of arrival to be reset to zero in edge cases.